refactor loginscreen
