--- app/screens/FriendsScreen.tsx
@@ -1,3 +1,0 @@
-import FriendsScreen from './friends/FriendsScreen';
-
-export default FriendsScreen;
